+++ imagepypelines/builtin_blocks/BlockViewer.py
@@ -32,7 +32,7 @@ class BlockViewer(SimpleBlock):
             training
         trained(bool): whether or not this block has been trained, True
             by default if requires_training = False
-        printer(ip.Printer): printer object for this block,
+        logger(ip.IpLogger): logger for this block,
             registered to 'name'
     """