+++ imagepypelines/builtin_blocks/Gray2Color.py
@@ -24,7 +24,7 @@ class Gray2Color(SimpleBlock):
             training
         trained(bool): whether or not this block has been trained, True
             by default if requires_training = False
-        printer(ip.Printer): printer object for this block,
+        logger(ip.IpLogger): logger for this block,
             registered to 'name'
     """
     def __init__(self):