+++ imagepypelines/builtin_blocks/IFFT.py
@@ -25,7 +25,7 @@ class IFFT(SimpleBlock):
             training
         trained(bool): whether or not this block has been trained, True
             by default if requires_training = False
-        printer(ip.Printer): printer object for this block,
+        logger(ip.IpLogger): logger for this block,
             registered to 'name'
 
     """