+++ imagepypelines/builtin_blocks/PCA.py
@@ -28,8 +28,8 @@ class PCA(BatchBlock):
         data = np.vstack(data)
         # checking to make sure that enough components are specified
         if data.shape[1] < self.n_components:
-            self.printer.warning("more components specified than features!")
-            self.printer.warning("truncating n_components({}) to num_features({})"\
+            self.logger.warning("more components specified than features!")
+            self.logger.warning("truncating n_components({}) to num_features({})"\
                                     .format(self.n_components,data.shape[1]))
             # reinstantiating the class with fewer components
             # this is done to make sure that self.io_map is accurate