+++ imagepypelines/core/util.py
@@ -4,7 +4,7 @@
 # @github: https://github.com/jmaggio14/imagepypelines
 #
 # Copyright (c) 2018-2019 Jeff Maggio, Nathan Dileas, Ryan Hartzell
-from .Printer import get_printer
+import logging
 from .constants import CV2_INTERPOLATION_TYPES
 from .constants import NUMPY_TYPES
 from .Exceptions import InvalidInterpolationType
@@ -13,7 +13,7 @@ import inspect
 import collections
 import time
 
-TIMER_PRINTER = get_printer('TIMER')
+TIMER_PRINTER = logging.getLogger('TIMER')
 
 ################################################################################
 #                                 Error Checking
@@ -69,7 +69,7 @@ def print_args(func):
         VARPOSITIONAL = 'var-positional|'
         VARKEYWORD    = 'var-keyword   |'
         DEFAULT       = 'default       |'
-        PRINTER = get_printer(func.__name__)
+        PRINTER = logging.getLogger(func.__name__)
 
         arg_dict = collections.OrderedDict()
         vtypes = {}